+++ java/src/org/openqa/selenium/remote/http/jdk/JdkHttpMessages.java
@@ -19,9 +19,11 @@ package org.openqa.selenium.remote.http.jdk;
 
 import org.openqa.selenium.remote.http.AddSeleniumUserAgent;
 import org.openqa.selenium.remote.http.ClientConfig;
+import org.openqa.selenium.remote.http.Contents;
 import org.openqa.selenium.remote.http.HttpRequest;
 import org.openqa.selenium.remote.http.HttpResponse;
 
+import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.net.URI;
 import java.net.URLEncoder;
@@ -68,11 +70,12 @@ class JdkHttpMessages {
         break;
 
       case POST:
-          builder = builder.POST(BodyPublishers.ofInputStream(req.getContent()));
-          break;
+        // Copy the content into a byte array to avoid reading the content inputstream multiple times.
+        builder = builder.POST(BodyPublishers.ofByteArray(Contents.bytes(req.getContent())));
+        break;
 
       case PUT:
-        builder = builder.PUT(BodyPublishers.ofInputStream(req.getContent()));
+        builder = builder.PUT(BodyPublishers.ofByteArray(Contents.bytes(req.getContent())));
         break;
 
       default:
@@ -80,6 +83,11 @@ class JdkHttpMessages {
     }
 
     for (String name : req.getHeaderNames()) {
+      // Avoid explicitly setting content-length
+      // This prevents the IllegalArgumentException that states 'restricted header name: "Content-Length"'
+      if (name.equalsIgnoreCase("content-length")) {
+        continue;
+      }
       for (String value : req.getHeaders(name)) {
         builder = builder.header(name, value);
       }
@@ -106,12 +114,17 @@ class JdkHttpMessages {
     return rawUrl;
   }
 
-  public HttpResponse createResponse(java.net.http.HttpResponse<InputStream> response) {
+  public URI getRawUri(HttpRequest req) {
+    String rawUrl = getRawUrl(config.baseUri(), req.getUri());
+    return URI.create(rawUrl);
+  }
+
+  public HttpResponse createResponse(java.net.http.HttpResponse<byte[]> response) {
     HttpResponse res = new HttpResponse();
     res.setStatus(response.statusCode());
     response.headers().map()
       .forEach((name, values) -> values.stream().filter(Objects::nonNull).forEach(value -> res.addHeader(name, value)));
-    res.setContent(response::body);
+    res.setContent(() -> new ByteArrayInputStream(response.body()));
 
     return res;
   }