+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -19,9 +19,13 @@
 
 * Set min node support to 14.x.x (issue 1 in #10970)
 * fix: geckodriver session with node 18.x.x (issue 2 in #10970)
+* fix: JS firefox driver crashes on setting a profile (fixed with [commit](https://github.com/SeleniumHQ/selenium/commit/fa6deeea6bda1e73317157845772e114bd569b7d))
+* fix: "SetExperimental" option is not available in webdriverjs (Javascript/Typescript) (#10959)
+* fix: Do not allow Select class to select disabled options (#10812)
 
 #### :nail_care: Polish
 
+* Stop sending desiredCapabilities to local end and remote end
 * tests: Quit driver instance after each test
 * tests: Adds select by multiple(index,value, text) tests
 * chore/code style changes, add util (#10975)
@@ -29,6 +33,8 @@
 * destructuring exports
 * Moving from static to const to make js-dossier happy
 * Removing circular dependency Between webdriver.js and http.js
+* fix some typos in code and documentation 
+* add cdp v105 remove v102
 
 ## v4.4.0
 * Add support CDP 104 and remove CDP 101